fix some minor points

+++ src/pretrain/utils/data.py
@@ -13,7 +13,7 @@ def random_crop_resize(sample, crop_minlen, input_size):
 
     # if sample is too short (less than crop_minlen)
     if sample.shape[1] < crop_minlen:
-        sample_padded = np.zeros((2, crop_minlen), dtype=np.float64)
+        sample_padded = np.zeros((2, crop_minlen), dtype=np.float32)
         sample_padded[:, :sample.shape[1]] = sample
         sample = sample_padded
